IGA Refactoring and minor bug fix: - Add kv api: Instruction::getSubfunction - Rename EncoderBase to Encoder - Constexpr fix - Fixes for 32b warnings

+++ visa/iga/IGALibrary/Backend/GED/GEDBitProcessor.cpp
@@ -37,7 +37,7 @@ GEDBitProcessor::GEDBitProcessor(
 }
 
 bool GEDBitProcessor::isAlign16MathMacroRegisterCsrOperand(
-    Operand::Kind opKind, RegName regName, uint8_t regNum) const
+    Operand::Kind opKind, RegName regName, uint16_t regNum) const
 {
     return
         isAlign16MathMacroRegisterCsrPlatform() &&
