+++ visa/iga/IGALibrary/Backend/GED/GEDBitProcessor.hpp
@@ -60,7 +60,7 @@ namespace iga
         bool isAlign16MathMacroRegisterCsrOperand(
             Operand::Kind opKind,
             RegName reg,
-            uint8_t regNum) const;
+            uint16_t regNum) const;
         bool isAlign16MathMacroRegisterCsrOperand(const Operand &op) const{
             return isAlign16MathMacroRegisterCsrOperand(
                 op.getKind(),
